Refactor main with new HelpService - s3i12
