+++ src/service/HelpService.kt
@@ -1,0 +1,25 @@
+package service
+
+class HelpService {
+    fun printHelp() {
+        """
+        Возможные аргументы программы:
+        -h                              - Вызов справки
+        
+        Аутентифицироваться 
+        -login  <string>                - Логин пользователя, строка, строчные буквы. Не более 10 символов
+        -pass   <string>                - Пароль пользователя
+        
+        Авторизоваться
+        -res    <PATH.TO.RESOURCE>      - Абсолютный путь до запрашиваемого ресурса
+        -role   <READ|WRITE|EXECUTE>    - Уровень доступа к ресурсу
+        
+        Активность
+        -ds     <yyyy-mm-dd>            - Дата начала сессии с ресурсом
+        -de     <yyyy-mm-dd>            - Дата окончания сессии с ресурсом, формат
+        -vol    <int>	                - Потребляемый объем, целое число
+    """.trimIndent().also {
+            println(it)
+        }
+    }
+}
